+++ fortinet-fortirecon-aci/aci_operations/stealers_on_sale.py
@@ -30,3 +30,12 @@ def get_stealers_infections_on_sale(config: dict, params: dict) -> dict:
         params["end_date"] = MK.handle_date(params.get("end_date"))
     response = MK.make_request(endpoint=endpoint, method="GET", params=params)
     return response
+
+
+def update_stealers_on_sale_status(config: dict, params: dict) -> dict:
+    MK = MakeRestApiCall(config=config)
+    status = params.pop("status")
+    payload = {"status": status}
+    endpoint = '/aci/{org_id}/stealers_infections/on_sale'+'/{id}'.format(id=params.pop('stealers_on_sale_id'))
+    response = MK.make_request(endpoint=endpoint, method="PATCH", params=params, json_data=payload)
+    return response